+++ corepy/arch/ppc/isa/ppc_fields.py
@@ -1,4 +1,5 @@
 # Copyright (c) 2006-2009 The Trustees of Indiana University.                   
+# Copyright (c) 2025 Max Wu EfiPy.Core@gmail.com
 # All rights reserved.                                                          
 #                                                                               
 # Redistribution and use in source and binary forms, with or without            
@@ -51,7 +52,7 @@ class PPCField(InstructionOperand):
     return
 
   def render(self, value):
-    return (long(value) & self.bit_mask)  << self.shift
+    return (int(value) & self.bit_mask)  << self.shift
 
 
 # TODO: Separate these out by type a little more like the x86 versions
@@ -65,7 +66,7 @@ class ConstantField(PPCField):
     return True
 
   def render(self, value):
-    return (long(self.const) & self.bit_mask)  << self.shift
+    return (int(self.const) & self.bit_mask)  << self.shift
 
 
 class RegisterField(PPCField):
@@ -79,19 +80,19 @@ class RegisterField(PPCField):
 
   def render(self, value):
     if isinstance(value, Register):
-      return (long(value.reg) & self.bit_mask) << self.shift
+      return (int(value.reg) & self.bit_mask) << self.shift
     elif isinstance(value, Variable):
-      return (long(value.reg.reg) & self.bit_mask) << self.shift
+      return (int(value.reg.reg) & self.bit_mask) << self.shift
     else:
-      return (long(value) & self.bit_mask) << self.shift
+      return (int(value) & self.bit_mask) << self.shift
 
 
 class SplitField(PPCField):
   def check(self, value):
-    return isinstance(value, (int, long)) and value >= 0 and value < (1 << 10)
+    return isinstance(value, (int, int)) and value >= 0 and value < (1 << 10)
 
   def render(self, value):  
-    return (((long(value) & 0x3E0) >> 5) | ((long(value) & 0x1F) << 5)) << self.shift
+    return (((int(value) & 0x3E0) >> 5) | ((int(value) & 0x1F) << 5)) << self.shift
 
 
 class ImmediateField(PPCField):
@@ -100,7 +101,7 @@ class ImmediateField(PPCField):
     PPCField.__init__(self, name, position, default = default)
 
   def check(self, value):
-    return isinstance(value, (int, long)) and self.range[0] <= value and value < self.range[1]
+    return isinstance(value, (int, int)) and self.range[0] <= value and value < self.range[1]
 
 
 class TruncatedField(ImmediateField):
@@ -114,7 +115,7 @@ class TruncatedField(ImmediateField):
     return
 
   def render(self, value):
-    return ((long(value) >> self.truncate)  & self.bit_mask)  << self.shift
+    return ((int(value) >> self.truncate)  & self.bit_mask)  << self.shift
 
 
 class LabelField(TruncatedField):
