+++ corepy/arch/vmx/isa/vmx_fields.py
@@ -1,4 +1,5 @@
 # Copyright (c) 2006-2009 The Trustees of Indiana University.                   
+# Copyright (c) 2025 Max Wu EfiPy.Core@gmail.com
 # All rights reserved.                                                          
 #                                                                               
 # Redistribution and use in source and binary forms, with or without            
@@ -51,7 +52,7 @@ class VMXField(InstructionOperand):
     return
 
   def render(self, value):
-    return (long(value) & self.bit_mask)  << self.shift
+    return (int(value) & self.bit_mask)  << self.shift
 
 
 class RegisterField(VMXField):
@@ -65,11 +66,11 @@ class RegisterField(VMXField):
 
   def render(self, value):
     if isinstance(value, Register):
-      return (long(value.reg) & self.bit_mask) << self.shift
+      return (int(value.reg) & self.bit_mask) << self.shift
     elif isinstance(value, Variable):
-      return (long(value.reg.reg) & self.bit_mask) << self.shift
+      return (int(value.reg.reg) & self.bit_mask) << self.shift
     else:
-      return (long(value) & self.bit_mask) << self.shift
+      return (int(value) & self.bit_mask) << self.shift
 
 
 class ImmediateField(VMXField):
@@ -78,7 +79,7 @@ class ImmediateField(VMXField):
     VMXField.__init__(self, name, position, default = default)
 
   def check(self, value):
-    return isinstance(value, (int, long)) and self.range[0] <= value and value < self.range[1]
+    return isinstance(value, (int, int)) and self.range[0] <= value and value < self.range[1]
 
 
 OPCD = VMXField("OPCD", (0, 5))
