+++ corepy/arch/x86_64/types/x86_64_types.py
@@ -1,4 +1,5 @@
 # Copyright (c) 2006-2009 The Trustees of Indiana University.                   
+# Copyright (c) 2025 Max Wu EfiPy.Core@gmail.com
 # All rights reserved.                                                          
 #                                                                               
 # Redistribution and use in source and binary forms, with or without            
@@ -91,7 +92,7 @@ def _most_specific(a, b, default = None):
     
   return default
   
-_int_literals = (spe.Immediate, int, long)
+_int_literals = (spe.Immediate, int, int)
 
 class PPCType(spe.Type):
   def _get_active_code(self):
@@ -107,7 +108,7 @@ class PPCType(spe.Type):
 
 class BitType(PPCType):
   register_type_id = 'gp'
-  literal_types = (int,long)
+  literal_types = (int,int)
 
   def _upcast(self, other, inst):
     return inst.ex(self, other, type_cls = _most_specific(self, other))
